Exercise 4.21

+++ part4/bloglist-backend/controllers/blogs.js
@@ -35,6 +35,18 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id)
+    return response.status(401).json({ error: 'token invalid' });
+
+  const userId = decodedToken.id.toString();
+
+  const blogToDelete = await Blog.findById(request.params.id);
+  if (!blogToDelete) return response.status(404).end();
+
+  if (blogToDelete.user.toString() !== userId)
+    return response.status(401).json({ error: 'deletion not authorized' });
+
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
